common tests files placement refactor
