+++ test/gtest/common/all.cpp
@@ -2,6 +2,8 @@
 //
 // SPDX-License-Identifier: BSD-3-Clause
 
+#include "xhp-tests.hpp"
+
 // Fixture
 template <typename T> class All : public testing::Test {
 public:
