+++ test/gtest/mhp/alignment.cpp
@@ -2,7 +2,7 @@
 //
 // SPDX-License-Identifier: BSD-3-Clause
 
-#include "mhp-tests.hpp"
+#include "xhp-tests.hpp"
 
 using DV = dr::mhp::distributed_vector<int>;
