+++ test/gtest/mhp/distributed_vector.cpp
@@ -2,7 +2,7 @@
 //
 // SPDX-License-Identifier: BSD-3-Clause
 
-#include "mhp-tests.hpp"
+#include "xhp-tests.hpp"
 
 using T = int;
 using DV = dr::mhp::distributed_vector<T>;
