--- test/gtest/mhp/mhp-subrange-tests.cpp
@@ -1,7 +1,0 @@
-// SPDX-FileCopyrightText: Intel Corporation
-//
-// SPDX-License-Identifier: BSD-3-Clause
-
-#include "mhp-tests.hpp"
-
-#include "common/subrange.hpp"
