+++ test/gtest/mhp/stencil.cpp
@@ -2,7 +2,7 @@
 //
 // SPDX-License-Identifier: BSD-3-Clause
 
-#include "mhp-tests.hpp"
+#include "xhp-tests.hpp"
 
 using T = int;
 using V = std::vector<T>;
