+++ test/gtest/shp/copy.hpp
@@ -2,7 +2,7 @@
 //
 // SPDX-License-Identifier: BSD-3-Clause
 #pragma once
-#include "shp-tests.hpp"
+#include "xhp-tests.hpp"
 
 template <typename AllocT> class CopyTest : public testing::Test {
 public:
