+++ test/gtest/shp/fill.cpp
@@ -1,7 +1,7 @@
 // SPDX-FileCopyrightText: Intel Corporation
 //
 // SPDX-License-Identifier: BSD-3-Clause
-#include "shp-tests.hpp"
+#include "xhp-tests.hpp"
 
 template <typename AllocT> class FillTest : public testing::Test {
 public:
