+++ test/gtest/shp/xhp-tests.hpp
@@ -9,8 +9,9 @@
 #include <fmt/ranges.h>
 #include <gtest/gtest.h>
 
-extern std::size_t comm_rank;
-extern std::size_t comm_size;
+// To share tests with MHP
+const std::size_t comm_rank = 0;
+const std::size_t comm_size = 1;
 
 // Namespace aliases and wrapper functions to make the tests uniform
 namespace xhp = dr::shp;
@@ -27,3 +28,6 @@ concept compliant_view = rng::forward_range<V> && requires(V &v) {
 };
 
 #include "common-tests.hpp"
+
+using AllTypes = ::testing::Types<xhp::distributed_vector<int>,
+                                  xhp::distributed_vector<float>>;
